fix: make `jkp connect` actually verify the WRDS connection
`jkp connect` only resolved credentials locally and printed
`Connected as: <user>`, so it reported success even when the
credentials were wrong/expired or WRDS was unreachable — users
only hit the failure later at `jkp build`.

Extract the WRDS connection primitives into a new
`wrds_connection.py` module and add `verify_wrds_connection()`,
which attaches WRDS read-only and runs a trivial query (a real
auth/connectivity/MFA check, bounded by connect_timeout). `connect`
now calls it before echoing success; failures surface a
password-free message and exit non-zero. `--reset` unchanged.

Keeping the primitives in a standalone module lets the CLI verify a
connection without importing the heavy pipeline module.

Closes #246
